requirements.txt
fastapi[standard]
python-dotenv
openai
psycopg[binary]
pgvector
pandas